maas-backend/requirements.txt
Django==5.2.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
python-decouple==3.8
django-cors-headers==4.3.1
PyMySQL==1.1.1
gunicorn==22.0.0
Pillow==10.4.0
